requirements.txt
pytorch
torchvision
torchaudio
google-generativeai
json
